refactor(quest): consolidate /quests into /quest (single router)
- Merge quests.py endpoints into quest.py under /quest prefix
- Remove quests.py (no longer needed)
- Rename test_quests.py -> test_quest_non_generate.py
- Update api.py: remove quests router, keep quest/analyze/users
- Keep test_quest.py (generate) separate by design
